Add a primitive pressure coupling.
At the moment this feature is supported only by the Landau free energy and CPU Euler integrator
