+++ src/integrators/EulerCPU.cpp
@@ -6,6 +6,12 @@ template<int dims>
 EulerCPU<dims>::EulerCPU(FreeEnergyModel *model, toml::table &config) : Integrator<dims>(model, config) {
     _N_per_dim_minus_one = this->_N_per_dim - 1;
 	_log2_N_per_dim = (int) std::log2(this->_N_per_dim);
+
+	_couple_pressure = this->template _config_optional_value<bool>(config, "couple_pressure", false);
+	if(_couple_pressure) {
+		_pressure_lambda = this->template _config_value<double>(config, "pressure_lambda");
+		_pressure_target = this->template _config_value<double>(config, "pressure_target");
+	}
 }
 
 template<int dims>
@@ -24,10 +30,22 @@ void EulerCPU<dims>::evolve() {
         }
     }
 
+	for(unsigned int idx = 0; idx < this->_N_bins; idx++) {
+        for(int species = 0; species < this->_model->N_species(); species++) {
+            rho_der(idx, species) = this->_M * _cell_laplacian(rho_der, species, idx);
+
+			if(_couple_pressure) {
+				Gradient<dims> rho_grad = this->_cell_gradient(this->_rho, species, idx);
+				double tot_pressure = this->_model->pressure(this->_rho.rho_species(idx)) - 0.5 * this->_k_laplacian * (rho_grad * rho_grad);
+				rho_der(idx, species) += _pressure_lambda * (tot_pressure - _pressure_target);
+			}
+        }
+    }
+
     // and then we integrate them
     for(unsigned int idx = 0; idx < this->_N_bins; idx++) {
         for(int species = 0; species < this->_model->N_species(); species++) {
-            this->_rho(idx, species) += this->_M * _cell_laplacian(rho_der, species, idx) * this->_dt;
+            this->_rho(idx, species) += rho_der(idx, species) * this->_dt;
         }
     }
 }
@@ -91,6 +109,57 @@ double EulerCPU<2>::_cell_laplacian(RhoMatrix<double> &field, int species, int i
 			field(_cell_idx(coords_xyp), species) -
 			4 * field(idx, species))
 			/ SQR(this->_dx);
+}
+
+template<>
+Gradient<1> EulerCPU<1>::_cell_gradient(RhoMatrix<double> &field, int species, int idx) {
+	int idx_p = (idx + 1) & _N_per_dim_minus_one;
+
+	return Gradient<1>({(field(idx_p, species) - field(idx, species)) / _dx});
+}
+
+template<>
+Gradient<2> EulerCPU<2>::_cell_gradient(RhoMatrix<double> &field, int species, int idx) {
+	int coords_xy[2];
+	_fill_coords(coords_xy, idx);
+
+	int coords_xpy[2] = {
+			(coords_xy[0] + 1) & _N_per_dim_minus_one,
+			coords_xy[1]
+	};
+
+	int coords_xyp[2] = {
+			coords_xy[0],
+			(coords_xy[1] + 1) & _N_per_dim_minus_one
+	};
+
+	return Gradient<2>({
+		(field(_cell_idx(coords_xpy), species) - field(_cell_idx(coords_xy), species)) / this->_dx, 
+		(field(_cell_idx(coords_xyp), species) - field(_cell_idx(coords_xy), species)) / this->_dx
+	});
+}
+
+template<>
+double EulerCPU<1>::_divergence(RhoMatrix<Gradient<1>> &flux, int species, int idx) {
+	int idx_m = (idx - 1 + _N_bins) & _N_per_dim_minus_one;
+	return (flux(idx, species)[0] - flux(idx_m, species)[0]) / this->_dx;
+}
+
+template<int dims>
+double EulerCPU<dims>::_divergence(RhoMatrix<Gradient<dims>> &flux, int species, int idx) {
+	double res = 0;
+	int coords[dims], coords_m[dims];
+	this->_fill_coords(coords, idx);
+	memcpy(coords_m, coords, sizeof(int) * dims);
+
+	for(int d = 0; d < dims; d++) {
+		coords_m[d] = (coords[d] - 1 + this->_N_bins) & this->_N_per_dim_minus_one;
+		int idx_m = this->_cell_idx(coords_m);
+		res += (flux(idx, species)[d] - flux(idx_m, species)[d]) / this->_dx;
+		coords_m[d] = coords[d];
+	}
+
+	return res;
 }
 
 template class EulerCPU<1>;
