+++ src/models/Landau.h
@@ -25,6 +25,7 @@ public:
 	void der_bulk_free_energy(field_type *psi, float *psi_der, int grid_size) override;
 	void der_bulk_free_energy(const RhoMatrix<double> &rho, RhoMatrix<double> &rho_der) override;
 	double bulk_free_energy(const std::vector<double> &) override;
+	double pressure(const std::vector<double> &) override;
 
 	GET_NAME(Landau free energy model)
